Refactor model key sanitizer and improve top deals queries
